feature: Create useClickOutside
