+++ src/components/Popover/Popover.jsx
@@ -1,29 +1,12 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useRef } from 'react';
 import PropTypes from 'prop-types';
 import * as Styled from './styles';
 import Arrow from '../Icons/Arrow';
+import { useClickOutside } from '../../hooks/hooks';
 
 const Popover = React.memo(({ popoverLabel, children }) => {
-  const [isOpen, toggleIsOpen] = useState(false);
   const node = useRef(null);
-
-  useEffect(() => {
-    const handleClickOutside = (e) => {
-      if (node.current && !node.current.contains(e.target)) {
-        toggleIsOpen(false);
-      }
-    };
-
-    document.addEventListener('click', handleClickOutside, true);
-
-    return () => {
-      document.removeEventListener('click', handleClickOutside, true);
-    };
-  });
-
-  const changeIsOpen = () => {
-    toggleIsOpen(true);
-  };
+  const [isOpen, changeIsOpen] = useClickOutside(node);
 
   return (
     <Styled.PopoverWrap>
